Change fhg type converter creation
